+++ prisma/migrations/20260128124741_add_phone_date_name_field_types/migration.sql
@@ -1,5 +1,0 @@
--- AlterEnum
--- Add new field types: PHONE, DATE, NAME
-ALTER TYPE "FieldType" ADD VALUE 'PHONE';
-ALTER TYPE "FieldType" ADD VALUE 'DATE';
-ALTER TYPE "FieldType" ADD VALUE 'NAME';